calculate fee and tax

+++ backend/app/Service/StockService.php
@@ -45,19 +45,19 @@ class StockService
         ]);
     }
 
-    public function calculateFee(int $price, int $quantity): int
+    public function calculateFee(float $price, int $quantity): int
     {
         $discount = $this->systemPreferenceRepository->getSystemPreferenceByName(SystemPreferenceService::BROKERAGE_FIRM);
         return ($price * $quantity * 0.001425 * floatval($discount)) > 1 ? intval($price * $quantity * 0.001425 * floatval($discount)) : 1;
     }
 
-    public function calculateTax(int $price, int $quantity, bool $isEtf = false): int
+    public function calculateTax(float $price, int $quantity, bool $isEtf = false): int
     {
         $proportion = $isEtf ? 0.001 : 0.003;
         return intval($price * $quantity * $proportion);
     }
 
-    public function calculateProfit(int $cost, int $price, int $quantity, bool $isEtf = false): int
+    public function calculateProfit(int $cost, float $price, int $quantity, bool $isEtf = false): int
     {
         $fee = $this->calculateFee($price, $quantity);
         $tax = $this->calculateTax($price, $quantity, $isEtf);
